requirements.txt
# Core dependencies
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
pandas==2.1.3
requests==2.31.0
Pillow==10.1.0
reportlab==4.0.7

# Optional: Web interface
streamlit==1.28.1

# Development dependencies
python-dotenv==1.0.0